requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
python-telegram-bot==20.7
streamlit
streamlit-authenticator
httpx
pytest
loguru
black
